New/STOC_/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
websockets==12.0
Babel==2.13.1
itsdangerous==2.1.2
python-jose==3.3.0
pywin32==306
